removed index page; added NEAR logo
