+++ styles/globals.scss
@@ -5,12 +5,17 @@ body {
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
+.logo {
+  max-width: 200px;
+}
+
 h1 {
   a {
     color: inherit;
     text-decoration: none;
   }
 }
+
 * {
   box-sizing: border-box;
 }
